+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add TUI switch for status and priority
 - Add default due config entry
 - Support numbered priority config
+- Allow set percent_complete and status for new todo
 
 ### Changed